+++ src/routes/twoColumn/useIsSecondColumn.ts
@@ -1,0 +1,8 @@
+import { createContext } from "react";
+import { use } from "react";
+
+export const IsSecondColumnContext = createContext<boolean>(false);
+
+export function useIsSecondColumn() {
+  return use(IsSecondColumnContext);
+}
